NXDRIVE-1673: Use native dialog when Qt is unavailable (for the fatal error screen only)
* On Windows, MessageBoxW() will be used.
* On macOs, osascript will be used.
* On GNU/Linux, our best effort will be to print to STDERR.
