+++ docs/changes/4.1.4.md
@@ -12,6 +12,7 @@ Release date: `2019-xx-xx`
 
 ## GUI
 
+- [NXDRIVE-1673](https://jira.nuxeo.com/browse/NXDRIVE-1673): Use native dialog when Qt is unavailable (for the fatal error screen only)
 - [NXDRIVE-1681](https://jira.nuxeo.com/browse/NXDRIVE-1681): Make server URL and local folder clickable in the Accounts settings
 - [NXDRIVE-1683](https://jira.nuxeo.com/browse/NXDRIVE-1683): Display a notification when DirectEdit'ing an inexistant document
 - [NXDRIVE-1684](https://jira.nuxeo.com/browse/NXDRIVE-1684): Display the file size in the systray
@@ -37,4 +38,8 @@ Release date: `2019-xx-xx`
 - Removed `BlackListItem.get()`. Use `name` attribute instead.
 - Added `BlacklistQueue.repush()`
 - Added `FileModel.ID` role
+- Moved \_\_main__.py::`check_executable_path` to fatal_error.py
+- Moved \_\_main__.py::`show_critical_error` to fatal_error.py
+- Removed \_\_main__.py::`section`
+- Added fatal_error.py
 - Added utils.py::`sizeof_fmt()`
